test-case: handle sof-logger errors in check-sof-logger.sh
Previously, we use the file size generated by sof-logger
to test if we have firmware log to export. But after a
logger change, sof-logger errors will output to generated
file, the size of the file is variant due to sof-logger
errors.
This patch redirects sof-logger errors to a temp file, and
search sof-logger and fw errors with grep.

Signed-off-by: Amery Song <chao.song@intel.com>