test-case: refine check-sof-logger.sh test case
Previously, if sof-logger only catchs the firmware log header, we will
fail the test case. Obviously, this is an overaction. So this patch relaxes
the firmware log check rule by allowing the log size less than 4K.

As this test case focuses on sof-logger function check, if firmware
error is found in the log, we just print out the error without failing
the test case.

Signed-off-by: Wu, BinX <binx.wu@intel.com>