test-case: refine verify-pcm-list.sh
As none of our platforms support multiple topologies,
we don't need to support multiple topologies in one
test case. If we do need to test multiple topologies,
run the test case multiple times.

This patch removes the iteration of multiple topologies,
and make it only supported one topology, this makes the
test case clean and tidy.

Signed-off-by: Amery Song <chao.song@intel.com>