tools: enable filtering for KWD pipeline
Previously, DMIC16k pipeline is blocked, and no test
case will run on it, because it tends to be used for
keyword detection, and test cases cannot handle it.

This patch adds support for KWD pipeline filtering,
thus enables test on DMIC16k pipeline which is not
used for keyword detection.

Default KWD pipeline is hidden from other pipelines,
to acquire KWD pipeline, run:
    sof-tplgreader.py -f "kwd" tplg-path

Signed-off-by: Amery Song <chao.song@intel.com>