fix pulseaudio re-start command line
When running a kmod-load-unload test locally on a Ubuntu 20.04
machine, it couldn't re-start pulseaudio, the reason was, that the
script was trying to start the command in the form '   pulseaudio -D'
i.e. with several spaces in front of the executable name and that
didn't work. Whereas just using the line as obtained from "ps"
output works.

Signed-off-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>