verify-sof-firmware-load: stop using dmesg; ring buffers can expire
Switch to journalctl --dmesg which is slower because it gets the log
from boot which means it works.

Also remove the useless and confusing eval and the useless [ ]
test. Switch one grep to grep -q to avoid generating a massive string
and cap the other two to just 12 lines (enough to see some repetition if
any)

This was tricky to reproduce because it required some indirection script
like run-all-tests.sh.

Good reminder to gradually get rid of dmesg everywhere.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>