case-lib/lib.sh: add new fake_kern_error()
This function adds a fake error to dmesg.

It is surprisingly easy to write a test that always succeeds, especially
in shell script and it has already happened a number of
times. Temporarily add this function to a test under development to make
sure it can actually report failures. Using this function is even more
critical for testing changes to the test framework and especially error
handling code.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>