logger: stop discarding all error messages and fix the cleanup code
Likely to avoid the shell printing a "Killed logger" message, lib.sh was
starting the logger with a "2>/dev/null" redirection. Here's a sample
list of logger errors that have been discarded by this redirection since
forever:

 error: Unable to open ldc file /etc/sof/sof-apl.ldc
 error: Unable to open in file /sys/kernel/debug/sof/trace
 error: Unable to open version file /sys/kernel/debug/sof/fw_version

Remove this 2>/dev/null redirection which was an insane price to pay for
making the logs "beautiful".

While not important, let's make the logs "beautiful" anyway because we
can: thanks to -SIGINT instead of SIGKILL. SIGKILL (-9) does not give a
any chance for the logger to exit cleanly, flus outputs etc. and is
only ever used as a last resort, see "man systemd.kill" etc.

Also log the logger command complete with parameters and the size of its
output at the end.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>