sof-dump-status: fix ACPI support
The ACPI support is currently problematic and only works if there is a
PCI device, which is true only on MinnowBoard and Chromebook Cyan used
for SOF CI, but false in general for all other BYT/CHT devices.

There is no reason why ACPI needs to be tied to PCI, devices can be
found in /sys/bus/acpi/devices/ directly. In specific cases, such as
MinnowBoard, multiple ACPI devices can be found and a loop is needed
to identify which device is enabled by platform firmware.

This patch does not change the structures but in follow-up patches we
should revisit the use of lists for PCI and ACPI devices when we
really only need a single structure for each.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>