kmod: remove legacy drivers if loaded
When legacy drivers are built and driver run-time selection, the
sof_remove.sh script fails since a number of modules are in use. These
drivers are actually not doing anything since their probe failed, but
the modules are still loaded.

Remove them explicitly so that the script correctly handles
dependencies.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>