[WIP][RFC] Test-case: Set PGAs to unity gain in script check-alsabat.sh
This patch adds find of volume controls for used playback and capture
devices and sets them to unity gain (0 dB). The alsabat test fails
if e.g. capture PGA is set to their maximum value that can be up
to +30 dB. The test sine wave distorts and causes test fail.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>