env-check.sh: show which file(s) lack exec permissions. Other cleanups
It was not very useful to tell the user that some files lack execute
permission without saying which ones. Especially not useful when the
files missing execute permission are irrelevant files like foo.sh~
backup files. This is also simplifies the code.

Also rephrase the error messages and offer a simpler chmod fix

Signed-off-by: Marc Herbert <marc.herbert@intel.com>