lib.sh: revise card number acquisition code
Auxiliary devices in SOF feature is introduced in
thesofproject/linux#2501, and it adds a new sof-probes
card in /proc/asound/cards, which will break current
test, because we cannot acquire the correct card ID.

This patch workarounds the issue by filtering out the
sof-probes card with grep. Even it gets merged, our
test is safe to run.

Signed-off-by: Amery Song <chao.song@intel.com>