multiple-pipeline-capture.sh: fix recent set -e related regression
Fixes 202507fe0c71 ("multiple-pipeline: fix set -e compatibility and add
it") that didn't pay attention to the fact that sof-process-state.sh
returns 2 when there is no process match and broke the test when there
are only aplay or only arecord commands but not both at the same time.

This regression was observed (among others) at:
 https://sof-ci.01.org/linuxpr/PR2565/build4887/devicetest/?model=APL_UP2_NOCODEC&testcase=multiple-pipeline-capture

Also fixes more minor set -e compatibility issues.

I missed the failure in
 https://sof-ci.01.org/softestpr/PR497/build372/devicetest/?model=APL_UP2_NOCODEC&testcase=multiple-pipeline-capture
because it was reported as "N/A" and not FAIL because Jenkins interprets
an exit status '2' as N/A. I should have paid more attention.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>