lib.sh: log git version of the tests at the start of every test
While there is generally in test results some version information about
firmware and kernel (except for force-pushes, merges with target
branch,... I digress) we had absolutely none about the version of
sof-test itself. This made it very difficult to monitor sof-test fixes
and other sof-test changes. This is also useful when someone shares the
log of a local test run.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>