multiple-pipeline: remove initial sleep and clarify second sleep
Having the same -w parameter for both this initial sleep and the test
duration made no sense: fixed by removing the first sleep.

It's not clear what the purpose of this first sleep was, it has been in
multiple-pipeline-capture/playback.sh since the dawn of time (= commit
39cd964ddc776c) For sure this first sleep was helping hide
"uninterruptible state" bug #472 by giving the DSP more time to wake up!

Signed-off-by: Marc Herbert <marc.herbert@intel.com>