kmod: unload snd_sof_intel_hda_common early, work around dep. changes
From pierre-louis.bossart@linux.intel.com in PR #589

PR https://github.com/thesofproject/linux/pull/2683 introduces a
dependency change. There is still one case there a module is used
differently before and after this PR, and renaming isn't really an
option.

This patch suggests a temporary workaround where we try to remove a
module (with errors disabled), and if it doesn't work we keep going.
In practice the snd_sof_hda_common module is made mostly of helper
routines and tables, without any PCI stuff, so there is a limited risk
of issues.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>