lib: pick up card ID only from the line that starts with a number
If DMI info is missing and the card long name is replaced with
the card name set in the machine driver, there will be 2 lines
like this:

 0 [sofhdadsp      ]: sof-hda-dsp - sof-hda-dsp
                      sof-hda-dsp

And the test would pick up the card ID from the second line and
result in a failure.

So, ignore lines that dont start with a number to avoid this.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>