check-suspend-resume: sleep 1s before rtcwake iteration
After some kernel update, we start to observe a timeout
issue on several JF devices. As we are not going to fix
the kernel, this workaround is applied.

Accroding to the solution in this patch, the device may
get into sleep before netcat can send out messages, this
maybe related to network issue in the kernel, or network
route is too long between JF and SH, etc. In this patch,
we add a 1-second sleep to wait a stable connection before
running rtcwake.

Workaround: #650

Signed-off-by: Chao Song <chao.song@linux.intel.com>