hijack.sh: make late logger config changes to nudge stuck DMA traces
There are bugs that cause DMA drops and the last few lines to get stuck
somewhere. Changing the logger configuration and waiting for 1 second
seems to be enough to nudge the system and force it to DMA a bit more
logs: hopefully all the logs relevant to the current test.

Running this logger configuration command twice makes it very easy to
observe the stuck lines bug: the "ipc" logs corresponding to this -F
command will appear _only once_ at the end of the slogger.txt DMA trace!

Signed-off-by: Marc Herbert <marc.herbert@intel.com>