test-speaker: add argument for loop count
The existed '-l' is for nloops of speaker-test, which is used to
configure the loop number of left/right .wav files in each playback, it
is not suitable for the loop count for the speaker-test stress.

Change the old speaker-test one to '-L' and add new '-l' for stress test
purpose, to make it consistent with stress test of other cases.

Signed-off-by: Keyon Jie <yang.jie@linux.intel.com>