sof-tplgreader: enable filtering a group of pipelines
Previous fix ce58aeda removes unclear RE for pipeline
filtering, but it also removes the ability to filter
a group of pipelines.

For example, pcm names for HDMI pipelines are HDMIn,
where n is a number, with previous RE, we are able
to filter out all HDMI pipelines with '~pcm:HDMI'.
With the fix ce58aeda, we can only filter out explicit
pipeline, for example HDMI1.

This patch uses full match for numerical values for exact
value match, and uses substring match for string values
for filtering a group of pipelines.

Signed-off-by: Chao Song <chao.song@linux.intel.com>