hijack.sh: print FAIL on missing logfile to avoid long CI timeout
For reasons beyond the scope of this commit, our Jenkins CI does not
know when a test script has finished to run. Instead of checking the
exit status like a normal test runner it waits for a "FAIL" or "PASS" or
"N/A" string to be printed. When the test aborts without printing one of
these, Jenkins falls back on a fairly long timeout - per test.

This unusual approach has been causing issues with `set -e` (#312)

In some relatively rare but possible failures the DSP or the drivers do
not load at all, examples:

https://sof-ci.01.org/sofpr/PR4766/build10353/devicetest/
https://sof-ci.01.org/sofpr/PR4768/build10348/devicetest/
https://sof-ci.01.org/sofpr/PR4782/build10387/devicetest/

In that case the $logFile is not found and many tests abort without
printing FAIL.

This commit prints FAIL when $logfile is missing and avoids the many
per-test Jenkins timeouts.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>