tplgtool2: dashed for dynamic pipelines
Previously, we need to check the topology source code to see if it has
dynamic pipelines. It is inconvient for the dynamic pipeline developers.
Here the topology tool will make the widgets under dynamic dashed and
blue in the diagram, so the developers can see whether there are dynamic
pipelines and which pipelines are dynamic.

Signed-off-by: Yongan Lu <yongan.lu@intel.com>