kernel-log-check: remove filter that ignores DSP boot retries
Now that sof-test is using journalctl log levels and that the levels of
this DSP retries have been fixed by
https://github.com/thesofproject/linux/pull/3250, it should not be
required to filter out these messages anymore. This cleanup removes
almost 30 lines of code.

This is a "logical" revert of the addition in d238c48cf4e6b and of the
many changes after it; most recently 61b077be7b28. See the code reviews
of all these commits for more details.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>