lib.sh: add 3 seconds to SOF_TEST_INTERVAL for our own delays
SOF_TEST_INTERVAL is the "sleep" delay between the end of one sof-test
and the start of the next one. However this does not account for our own
delays collecting logs, which makes some bugs like
https://github.com/thesofproject/sof/issues/5032 being caught
semi-randomly. Err on the safe side and add 3 seconds for more
determinism (this will of course never be completely deterministic).

Signed-off-by: Marc Herbert <marc.herbert@intel.com>