hijack.sh: read from -i /dev/null when changing the log level
sof-logger inconveniently does not allow changing the log level without
reading some logs at the same time. We've always redirected these logs
to /dev/null but Zephyr uses a different etrace format that can cause
errors on stderr and we don't want to discard stderr because we never
want to ignore ALL errors in test code. We can't switch to the DMA trace
either because reading the DMA trace never stops.

So, read the trace from neither the DMA nor the shared memory but from
/dev/null instead.

Keep redirecting output to /dev/null too because we still don't want to
see the banner either.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>