volume-basic-test: Change the control search expression
IPC4 uses "gain" as the widget name instead of "PGA". So searching for
"PGA" would result in the test getting skipped with IPC4. But both IPC3
and IPC4 have the pga widget names suffixed with
".Pipeline_ID.Instance_ID". So use the regular expression to search for
volume controls instead of searching for "PGA".

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>