check-kmod-load-unload: wait 10 seconds for firmware to boot
Now that the test has been just fixed in commit 71a9ccac4ecfa, of course
it started failing:

https://sof-ci.01.org/sofpr/PR5383/build12069/devicetest/?model=BYT_MB_NOCODEC&testcase=check-kmod-load-unload-after-playback

This proves the older version wasn't testing anything: classic case of
"green failure"
https://github.com/thesofproject/sof-test/issues?q=+label%3A%22False+Pass+%2F+green+failure%22+

Firmware does not boot instantly. We even have retries for when it
fails:

https://github.com/thesofproject/linux/pull/3255
https://github.com/thesofproject/linux/commit/2a3c9269474c

So, poll journalctl for a maximum of 10 seconds before failing.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>