kmod: add support for ES83x6 platforms
Add missing machine driver and codecs.

The codec names are confusing.

The snd_soc_es8316 codec driver can handle both the es8316 and es8336
skews.

The ACPI HID for APL+ platform is ESSX8336, so we wrote the sof_es8336
machine driver.

A new codec is used in some platforms, it will required the use of the
snd_soc_es8326 module, but the same machine driver will be used.

So to recap, we need to deal with two codec drivers and one common
machine driver.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>