verify-sof-firmware-load: print wallclocks timestamps too
Both ktimes and wallclock timestamps are useful but using sometimes the former
and other times the latter can make troubleshooting failures more
difficult. In this case print both. It is especially useful here
considering this test PASSes after unloading the driver.

Also update the -help message that was out of date with recent changes
and move logs after journalctl status in case of failure.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>