multiple-pipeline: skip Echo Reference pipeline when getting count
Previously, the Echo Reference pipeline has been skipped when running
the parallel pipeline test, but it wasn't skipped when calculating
PIPELINE_COUNT. Therefore, on all pipelines mode, the expected process
count is inconsistent with the actual running process. To solve this
issue, skip the same pipeline when acquiring the pipeline count for
consistency.

Signed-off-by: Yongan Lu <yongan.lu@intel.com>