suspend-resume: fix sys/power/wakeup_count test
"greater or equal" has never ever reported any error. Test for the
expected value instead but print only a warning for now because of the
high number of failures this causes, example:

  https://sof-ci.01.org/softestpr/PR900/build18/devicetest/

Also catch any unexpected increase while the system is up - trust no
one.

Fix the error message to report exactly what was found, not a guess.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>