sof-dump-status.py: Add read helper function
Since there's no point in creating a subprocess just to
read the content of a file, a new helper function was
added. The purpose of this function is to simply read
said content and return it.

The possible exceptions that may arrise will have to be
treated by the caller functions.

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>