Enable mtrace (#956)
* lib.sh: move logger collection to a function

Signed-off-by: Keqiao Zhang <keqiao.zhang@intel.com>

* lib.sh: add a function for mtrace collection

mtrace is a new logging tool. Tool to stream data from Linux
SOF driver mtrace debugfs interface to standard output.

Signed-off-by: Keqiao Zhang <keqiao.zhang@intel.com>

* lib.sh: rename the ipc4_used to is_ipc4

Align the name with other similar functions and
refine the is_ipc4 function.

Signed-off-by: Keqiao Zhang <keqiao.zhang@intel.com>

* lib.sh: add a function to check if the firmware file is built with zephyr

This function can check if the running firmware file is built with Zephyr.

Signed-off-by: Keqiao Zhang <keqiao.zhang@intel.com>

* lib.sh: add a function to check if the firmware file is built with zephyr

This function can check if the running firmware file is built with Zephyr.

Signed-off-by: Keqiao Zhang <keqiao.zhang@intel.com>

* lib.sh: enable the mtrace for IPC4 Zephyr platforms

The mtrace is only support IPC4 platforms and mtrace should run
while tests are running ATM. If mtrace is not running, DSP will
stop outputting logs when buffer is full. So we need to keep the
mtrace running during the test, or you won't see the error log.

Signed-off-by: Keqiao Zhang <keqiao.zhang@intel.com>

* hijack.sh: kill the mtrace after test

need to kill the mtrace-reader.py process after the test
to prevent the conflicts between each case.

Signed-off-by: Keqiao Zhang <keqiao.zhang@intel.com>

Signed-off-by: Keqiao Zhang <keqiao.zhang@intel.com>