tplgtool2.py: support filesystem globbing and multiple arguments
Warning: this is a backwards-incompatible change! The previous
behavior is now available ONLY when using `--tplgroot`.

Up to now the user interface made the unusual choice to implement
globbing internally instead of relying on the operating system. Maybe
for compatibility with ancient .BAT scripts? So before this commit,
internal globbing had to be quoted like this to escape the shell
globbing, otherwise it would reject too many arguments:

   ./tplgtool2.py 'sof-*.tplg'             # old, now INVALID usage

This commit adds support for multiple arguments and such quoting is not
necessary anymore. HOWEVER, for backwards-compatibility the old behavior
is preserved when using `--tplgroot`:

   ./tplgtool2.py --tplgroot . 'sof-*.tplg' # UNCHANGED behavior

See new --help message for more details.

Also add a number of `assert` to fix usability issue #969 and error
handling when relying on the older behavior.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>