Tools: Add more robustness to check_volume_levels.m measure
This patch delays the check window if the gain settling seems
to be still happening in the beginning of level check window.
The settling is detected with differential of gain.

The window means acceptance criteria over time as gain in
decibels. E.g. x milliseconds to y milliseconds, measured
gain within a +/- b decibels. The average gain from the measure
window is computed and compare to required. Any fail will
fail the entire test.

The differential of measured gain shows if the gain is constant
or changing. The added code is shifting the measurement window
to later in time if the gain is not constant in in the beginning.

The delay in gain settling can be caused by random alsamixer
controls apply delay. Also a very long gain ramp makes the settling
to appear later.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>