test-case: check-alsabat.sh: Lower the threshold for detecting tones
The default parameter sigma k is 3.0 when omitted from command line.
This patch drops the threshold to half with 1.5 to better detect
distortion in playback. Alsabat judges spectral components above the
threshold as sine frequencies.

This change should increase the probability fail a test to avoid
green failures to happen.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>