Revert "tools: DSM and ECHO REFERENCE pipelines filtering" partially
The multiple stream feature in IPC4 makes the previous
code for finding smart_amp/echo_reference playback and
capture hosts not work, because there are multiple
playback and capture hosts. Revert the code now, and
re-do a compatible version for both IPC3 and IPC4 later.

The smart_amp and echo_reference tests are not added in
CI, so this change won't break anything.

Signed-off-by: Chao Song <chao.song@linux.intel.com>