tplgtool2.py: Add more info info to the nodes in picture
It is sometimes hard to draw connection between the nodes in the
produced picture and user visible entities, like ALSA mixers and PCMs.
The old behavior is behind '-w' or '--without_nodeinfo' flags.

The copier nodes are added with, copier type: AIF_IN, AIF_OUT, DAI_IN
or DAI_OUT, and with copier widgets stream name.

Signed-off-by: Jyri Sarha <jyri.sarha@linux.intel.com>