lib.sh: make MAX_WAIT_FW_LOADING platform dependent
We don't want to wait for 70s in every test on every platform when the
firmware does not load. When multiplied by the number of tests in a plan
this can cause a considerable waste of hardware resources.

i915 timeout aside, 10s has proved over a long period to be more than
enough.

Fixes: #1059

Signed-off-by: Marc Herbert <marc.herbert@intel.com>