tplgtool2.py: fix --show_core handling of the lack of coreID
Fixes --show_core display option issues initially reported by Yong Zhi
in #1082

Components that don't have any core ID specified and using a default
core were mishandled in several ways:

- Default values did not count in the `is_multicore()` function! So,
  when using a single CORE_ID=2 the count was wrong and `is_multicore()`
  wrongly returned false. See #1082

- Default values were never shown even when using `--show_core always`

- It was impossible to show the difference between the default coreID (no
  value) and the explicit value zero.

Fix all these issues by giving the default value a real status in the
code.

A side-effect of this commit is that the `--show_core auto` option gets
rid of the "multicore" idea and will now display coreIDs when there is a
mix of "default" and "zero". This is a good thing. tplgtool2.py is about
analyzing and visualizing what is in .tplg files exactly. It shouldn't
be a tool that tries (and fails) to be smart and to guess whether a
topology is multicore. It shouldn't try to hide the difference between
no coreID and coreID zero.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>