Test-case: test-speaker.sh: Use pink noise for non-48 kHz rates
The speaker test fails with -t wav internally for other than
"-r 48000" because it is not re-sampling the wav "Front center"
etc. clips. Using pink noise for non-48 kHz works around this
limitation and allows to test with all sample rates. Also the
test are not depending on spoken channels names. Channels order
can be subjectively checked with the 48 kHz test.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>