sof_perf_analyzer: add skip-to-first-trace option
In CI test, some traces from previous test case will
apprear in the mtrace of current test case, this is
a know issue.

Previously, we always skip to the first valid trace
of this test case, this could raise trouble for developer
use of this script.

This patch adds skip-to-first-trace option, for CI
test, this option should be set (True), and for developer
use, this option should be left as it is (False).

Signed-off-by: Chao Song <chao.song@linux.intel.com>