check-sof-logger.sh: fix no-log-file failures due to RPM
The logger test case has a design limitation in that it
assumes the SOF DSP to remain powered up between driver reload
and start of the logger. This creates a possibility for a race
with runtime PM (RPM).

If the DSP goes to runtime suspend during this window, it is not
guaranteed that the SOF driver will provide any logs. Some logging
mechanisms save the logs that were emitted before suspend
(sof-logger/IPC3), but some not (cavstool/IPC3 and mtrace/IPC4).

To limit the window , add a simple ping of all PCM nodes just
before logger process is started. This is a crude mechanism but
something that works with all variants of SOF. End result is that
DSP will be resumed to D0 just before the logger test starts.

Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>