sof_perf_analyzer: introduce pandas for performance analysis
Previously, Python built-in structures and operations are
used for component performance data storage and calculation.
With Python built-in operations, the calculation is very
trivial, because only very few scalar and list operations
are supported.

Pandas is high-performance package for data analysis, it supports
element-wise, column-wise, row-wise and table-wise operation,
which is a good fit for our component performance analysis.

This patch changes the underlining python built-in data structure
used previously with pandas DataFrame and Series, while keeps the
program behavior unchanged.

As you can see from the diff, pandas does help us to simplify
the code a lot.

Signed-off-by: Chao Song <chao.song@linux.intel.com>