multiple-pause-resume.sh: Check for aplay/arecord pid when waiting
It has been observed that the expect process has been terminated (pidof
returns with no PID) but the child aplay is still running.
This only happens with ChainDMA enabled PCMs where the time to stop
the stream takes more than 100ms (from trigger stop to PCM close).
With the right timing on such PCM the pidof expect will break the wait
loop and we progress to the next iteration which - again with right
timing and setup - might fail if the aplay is still in a process of closing
and the next iteration includes the same PCM.

Add the aplay and arecord commands to the pidof command to make sure that
the iteration has been completed before starting a new one.
We will still catch timeouts and in that case we  should kill the stuck
processes.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>