tools: sof_perf_analyzer: use itertools.pairwise in kmsg parsing
In kernel message parsing, we need to process two adjacent
kernel messages to extract component name, component ID and
pipeline ID, itertools.pairwise is a good tool to provide
those pairs of kernel messages pair by pair.

itertools.pairwise helps to avoid the call of next() in iterating
an iterable, which is error-prone.

Signed-off-by: Chao Song <chao.song@linux.intel.com>