lib.sh: Workaround sof-test output alignment issue
If sof-test test case is run manually and locally,
the output of sof-test is messy, the lines are not
aligned left at the beginning of each line, some
lines start in the middle of a line, which makes
the output log hard to follow.

This patch helps to workaround the above issue.

Link: https://github.com/thesofproject/sof-test/issues/1151

Signed-off-by: Chao Song <chao.song@linux.intel.com>