kmod: sof_remove: Drop Pulseaudio test and stop including case-lib/lib.sh
The PA test does not make much sense anymore as distros are moving or moved
to use PW and the audio servers should have been disabled in the DUT
anyways to not interfere with tests (PA/PW will probe the new card on
module load for example).

Including case-lib/lib.sh have the side effect that it will 'start a test'.
This involves checking if the firmware is booted, but if we have booted up
in legacy or DSPless mode the script aborts and one has to figure out [1]
how to force the script to do what it supposed to be doing: removing the
modules.

This is a helper script, not a test case.

[1]
NO_POLL_FW_LOADING=false tools/kmod/sof_remove.sh
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>